Segregate client independent logic from validators
The WebAuthn specification is an application utilizes secure
authenticators which generates attestation and assertion.
This framework can be used not only for WebAuthn but also
for other purposes, and Apple App Attest is an application.

This commit segregates client, which means application, independent
core validation logic and make room to be reused by other applications
like App App Attest validators.