add support for additional BE and BS flags
As per the documentation https://w3c.github.io/webauthn/#authenticator-data , there are two additional flags available in authData to indicate backupEligbility and backupState for the authenticators. The flags essentially help us to resolve if an authenticator is a multi-device credential and has the credentials backedup. This change exposes the method to read the bits for the BE and BS flags.